+++ src/images/trufflesuite_ganachecli.rs
@@ -1,4 +1,3 @@
-use crate::core::Port;
 use crate::{Container, Docker, Image, WaitForMessage};
 use std::collections::HashMap;
 
@@ -6,7 +5,6 @@ use std::collections::HashMap;
 pub struct GanacheCli {
     tag: String,
     arguments: GanacheCliArgs,
-    ports: Option<Vec<Port>>,
 }
 
 #[derive(Debug, Clone)]
@@ -21,7 +19,6 @@ impl Default for GanacheCli {
         GanacheCli {
             tag: "v6.1.3".into(),
             arguments: GanacheCliArgs::default(),
-            ports: None,
         }
     }
 }
@@ -87,20 +84,7 @@ impl Image for GanacheCli {
         HashMap::new()
     }
 
-    fn ports(&self) -> Option<Vec<Port>> {
-        self.ports.clone()
-    }
-
     fn with_args(self, arguments: <Self as Image>::Args) -> Self {
         GanacheCli { arguments, ..self }
-    }
-}
-
-impl GanacheCli {
-    pub fn with_mapped_port<P: Into<Port>>(mut self, port: P) -> Self {
-        let mut ports = self.ports.unwrap_or_default();
-        ports.push(port.into());
-        self.ports = Some(ports);
-        self
     }
 }