+++ tests/cli_client.rs
@@ -4,7 +4,6 @@ use std::time::Instant;
 
 use spectral::prelude::*;
 
-use testcontainers::core::Port;
 use testcontainers::*;
 
 #[derive(Default)]
@@ -38,10 +37,6 @@ impl Image for HelloWorld {
 
     fn env_vars(&self) -> Self::EnvVars {
         HashMap::new()
-    }
-
-    fn ports(&self) -> Option<Vec<Port>> {
-        None
     }
 
     fn with_args(self, _arguments: <Self as Image>::Args) -> Self {